Separate assembly resolve in startup - netcore
